requirements.txt
# Core HTTP
httpx==0.27.0
tenacity==8.2.3

# Supabase
supabase==2.4.0

# AI
anthropic==0.28.0
openai==1.30.0

# Auth
PyJWT==2.8.0
cryptography==42.0.8

# Scheduling
APScheduler==3.10.4

# Web framework (FastAPI webhook listener for Airtable integration)
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Data
pandas==2.2.2
rapidfuzz==3.9.1
python-dateutil==2.9.0

# Utils
python-dotenv==1.0.1
structlog==24.1.0
pydantic==2.7.1
click==8.1.7
tqdm==4.66.4

# Web scraping (NYS DOS fallback)
beautifulsoup4==4.12.3
lxml==5.2.1

# Image processing (ACRIS TIFF → PDF conversion)
Pillow==10.3.0
